Initial work to setup processes and the scheduler (#11)
This change sets up the kernel to execute ring 3 processes in x86 64. As part of the changes in the pull request, I had to refactor the x86-64 dependent code into its own crate and only made public those things that should be controllable from the kernel context. This helps disinfect the kernel codebase from architecture-dependent code.

On top of that, I moved all the creates into a cargo workspace to guarantee a shared lock file.