Setup the scheduler with kernel threads. (#13)
This change sets up the kernel scheduler with kernel threads. At the moment the scheduler is able to preemptively switch between threads, however, there's currently no way to block a thread. Additionally, there seems to be a race condition when we switch from the main thread the first time. This can probably be improved.