Remove critical section
The kernel will generally be non-preemptible and lock free, so we don't
want to have a critical section. Right now things are likely safe (because we
are single core and non-preemptible), but in the future all of the static
muts will need to be akin to a static PerCore<T> or such that will provide mutable
references to a single core of execution.