Add shared user/kernel kapi interface
Kapi is the low level interface to perform syscalls somewhat coherently.

The goal of kapi is not to provide an easy-to-use abstraction but instead
to share the low level structures for interpreting syscalls. These
include CapIds, CapErrors, Operations, etc. This does not include
the actual use of a capability such as linking page tables or
modifying a capabiliy table. These operations will be available as part
of the librs which provides the helpers to perform low-level capability
manipulation.